requirements.txt
requests==2.32.3
azure-storage-blob==12.24.1
python-dotenv==1.0.1
selenium==4.28.1